+++ ft_strnstr.c
@@ -6,7 +6,7 @@
 /*   By: mrahmat- <mrahmat-@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/19 10:28:25 by mrahmat-          #+#    #+#             */
-/*   Updated: 2024/05/08 10:32:05 by mrahmat-         ###   ########.fr       */
+/*   Updated: 2024/05/10 10:51:57 by mrahmat-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char	*ft_strnstr(const char *haystack, const char *needle, size_t len)
 {
-	unsigned int	ni;
-	unsigned int	hi;
+	size_t	ni;
+	size_t	hi;
 
 	if (needle[0] == '\0')
 		return ((char *)haystack);
